requirements.txt
websockets
fastapi
openai
Pillow
python-dotenv
